refactor(util): extract `parse_time_span`
So it can be reused by the upcoming `-Zmin-publish-age`.
