+++ src/cargo/core/gc.rs
@@ -22,6 +22,7 @@
 use crate::core::global_cache_tracker::{self, GlobalCacheTracker};
 use crate::ops::CleanContext;
 use crate::util::cache_lock::{CacheLock, CacheLockMode};
+use crate::util::time_span::maybe_parse_time_span;
 use crate::{CargoResult, GlobalContext};
 use anyhow::{Context as _, format_err};
 use serde::Deserialize;
@@ -372,41 +373,6 @@ fn parse_time_span_for_config(config_name: &str, span: &str) -> CargoResult<Dura
     })
 }
 
-/// Parses a time span string.
-///
-/// Returns None if the value is not valid. See [`parse_time_span`] if you
-/// need a variant that generates an error message.
-fn maybe_parse_time_span(span: &str) -> Option<Duration> {
-    let Some(right_i) = span.find(|c: char| !c.is_ascii_digit()) else {
-        return None;
-    };
-    let (left, mut right) = span.split_at(right_i);
-    if right.starts_with(' ') {
-        right = &right[1..];
-    }
-    let count: u64 = left.parse().ok()?;
-    let factor = match right {
-        "second" | "seconds" => 1,
-        "minute" | "minutes" => 60,
-        "hour" | "hours" => 60 * 60,
-        "day" | "days" => 24 * 60 * 60,
-        "week" | "weeks" => 7 * 24 * 60 * 60,
-        "month" | "months" => 2_629_746, // average is 30.436875 days
-        _ => return None,
-    };
-    Some(Duration::from_secs(factor * count))
-}
-
-/// Parses a time span string.
-pub fn parse_time_span(span: &str) -> CargoResult<Duration> {
-    maybe_parse_time_span(span).ok_or_else(|| {
-        format_err!(
-            "expected a value of the form \
-             \"N seconds/minutes/days/weeks/months\", got: {span:?}"
-        )
-    })
-}
-
 /// Parses a file size using metric or IEC units.
 pub fn parse_human_size(input: &str) -> CargoResult<u64> {
     let re = regex::Regex::new(r"(?i)^([0-9]+(\.[0-9])?) ?(b|kb|mb|gb|kib|mib|gib)?$").unwrap();
@@ -441,6 +407,7 @@ pub fn parse_human_size(input: &str) -> CargoResult<u64> {
 
 #[cfg(test)]
 mod tests {
+    use crate::util::time_span::maybe_parse_time_span;
     use super::*;
     #[test]
     fn time_spans() {
